+++ test/test_all_streams.py
@@ -25,6 +25,7 @@ class ZendeskAllStreams(ZendeskTest):
             "ticket_fields",
             "group_memberships",
             "macros",
+            "satisfaction_ratings",
             "tags",
             "ticket_metrics",
         }
@@ -40,6 +41,7 @@ class ZendeskAllStreams(ZendeskTest):
             "ticket_fields": {"id"},
             "group_memberships": {"id"},
             "macros": {"id"},
+            "satisfaction_ratings": {"id"},
             "tags": {"name"},
             "ticket_metrics": {"id"},
         }
@@ -50,7 +52,7 @@ class ZendeskAllStreams(ZendeskTest):
         creds = {
             'email': 'dev@stitchdata.com',
             'password': os.getenv('TAP_ZENDESK_API_PASSWORD'),
-            'subdomain': "rjmdev",
+            'subdomain': self.get_properties()['subdomain'],
         }
 
         test_tags = ['test_tag_1', 'test_tag_2', 'test_tag_3']
@@ -72,6 +74,43 @@ class ZendeskAllStreams(ZendeskTest):
             zenpy_client.tickets.add_tags(tic.get('data').get('id'), test_tags[0:(i+1)])
             # mark tags as refreshed as soon as we successfully get through one loop
             self.tags_are_stale = False
+
+    def rate_tickets(self, records):
+        ''' WIP to collect and set satisfaction_ratings during test execution '''
+
+        # Zenpy client credentials to connect to API
+        creds = {
+            'email': 'dev@stitchdata.com',
+            'password': os.getenv('TAP_ZENDESK_API_PASSWORD'),
+            'subdomain': self.get_properties()['subdomain'],
+        }
+
+        ratings = ['bad', 'badwithcomment', 'good', 'goodwithcomment', 'offered']
+        closed_tickets = [t for t in records.get('tickets').get('messages') if t.get('data').get('status') == 'closed']
+        self.assertGreaterEqual(len(closed_tickets), 9)
+        last_9_closed_tickets = closed_tickets[-9:]
+        confirmed_rated_tickets = 0
+
+        zenpy_client = Zenpy(**creds)
+
+        for i, tic in enumerate(last_9_closed_tickets):
+            # if rating is not defaulted to unoffered then set it, otherwise skip
+            if tic.get('data').get('satisfaction_rating') != {'score': 'unoffered'}:
+                print(f"Skipping non-default rating! {tic.get('data').get('satisfaction_rating')}")
+                continue
+
+            # TODO move from client to curl or request as client generates the following errors?
+            # ipdb> zenpy_client.tickets.rate(2297, 'good')
+            # *** AttributeError: 'PrimaryEndpoint' object has no attribute 'satisfaction_ratings'
+            # ipdb> zenpy_client.tickets.rate(tic.get('data').get('id'), ratings[i%5])
+            # *** AttributeError: 'PrimaryEndpoint' object has no attribute 'satisfaction_ratings'
+            # ipdb> zenpy_client.tickets.rate(tic.get('data').get('id'), {'score': 'good'})
+            # *** AttributeError: 'PrimaryEndpoint' object has no attribute 'satisfaction_ratings'
+            # ipdb> zenpy_client.tickets.rate(2301, {"satisfaction_rating": {"score": "good", "comment": "Awesome support. Test 1"}})
+            # *** AttributeError: 'PrimaryEndpoint' object has no attribute 'satisfaction_ratings'
+
+            zenpy_client.tickets.rate(tic.get('data').get('id'), {'score': ratings[i%5]})
+            #zenpy_client.tickets.rate(id, rating) # example rating {'score': 'good'}
 
 
     def test_run(self):
@@ -108,10 +147,11 @@ class ZendeskAllStreams(ZendeskTest):
         # Verify exit status is 0 and verify rows were synced
         _ = self.run_and_verify_sync(conn_id, state={})
 
-        # Verify actual rows were synced
-
         # Ensure all records have a value for PK(s)
         records = runner.get_records_from_target_output()
+
+        # Ensure tickets data have some ratings now that we have records to check
+        # self.rate_tickets(records)  # TODO tickets.rate(id, rating) fails with client.
 
         # assume tags are stale since we cannot query tag age / date from synced records or the API
         self.tags_are_stale = True
@@ -142,7 +182,7 @@ class ZendeskAllStreams(ZendeskTest):
                     # tags were already refreshed so records were missing from first sync
                     messages = tags_records.get(stream).get('messages')
 
-            if stream in  ['tickets', 'groups', 'users']:
+            if stream in  ['groups', 'organizations', 'tickets', 'users']:
                 self.assertGreater(len(messages), 100, msg="Stream {} has fewer than 100 records synced".format(stream))
             for m in messages:
                 pk_set = self.expected_pks()[stream]